Code cleanup

+++ build.sh
@@ -5,7 +5,7 @@ BINDIR="bin"
 DATASYMLINK="${BINDIR}/data"
 
 CC="clang"
-CFLAGS="-c -g -DMACOSX -Isrc"
+CFLAGS="-c -Wall -Wno-missing-braces -g -DMACOSX -Isrc"
 LIBS="-framework Cocoa -framework OpenGL"
 
 function prepare() {
@@ -26,11 +26,17 @@ function build_renderer() {
   prepare
 
   $CC src/renderer/renderer.cpp $CFLAGS -o $OBJDIR/renderer.o
-  libtool -macosx_version_min 10.11 -dynamic $OBJDIR/renderer.o -lSystem -o $OBJDIR/renderer.dylib
   result=$?
-  cp $OBJDIR/renderer.dylib $BINDIR/renderer.dylib
 
-  killall -USR1 evolve
+  if [ $result -eq 0 ]; then
+    libtool -macosx_version_min 10.11 -dynamic $OBJDIR/renderer.o -lSystem -o $OBJDIR/renderer.dylib
+    result=$?
+    cp $OBJDIR/renderer.dylib $BINDIR/renderer.dylib
+  fi
+
+  if [ $result -eq 0 ]; then
+    killall -USR1 evolve
+  fi
 }
 
 function build_exe() {
